refactor: share benchmark code with simple flow test
