+++ crates/blockifier/Cargo.toml
@@ -10,7 +10,7 @@ description = "The transaction-executing component in the Starknet sequencer."
 workspace = true
 
 [features]
-testing = ["rstest"]
+testing = ["rand", "rstest"]
 concurrency = []
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -39,6 +39,7 @@ num-traits.workspace = true
 num-rational.workspace = true
 once_cell.workspace = true
 phf.workspace = true
+rand = { workspace = true, optional = true }
 rstest = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true, features = ["arbitrary_precision"] }
