+++ crates/blockifier/src/blockifier.rs
@@ -2,3 +2,5 @@ pub mod block;
 pub mod config;
 pub mod stateful_validator;
 pub mod transaction_executor;
+#[cfg(test)]
+pub mod transfers_flow_test;
